backend/requirements.txt
Django==2.1.7
pytz==2018.9

djangorestframework==3.9.2
django-cors-headers==2.5.2

djangorestframework-jwt==1.11.0

# DB drivers
psycopg2==2.7.7
psycopg2-binary==2.7.7

# # Swagger for API documentation
django-rest-swagger==2.2.0

django-debug-toolbar==1.11